1. Changed directory name 2. added .travis.yml file
