+++ .travis.yml
@@ -1,0 +1,5 @@
+language: java
+install: 
+  - mvn install
+jdk:
+  - oraclejdk8
