+++ src/main/java/cs/dal/csci3130/Undined/RequestForm.java
@@ -1,4 +1,4 @@
-package my.vaadin.Undined;
+package cs.dal.csci3130.Undined;
 
 import com.vaadin.event.ShortcutAction;
 import com.vaadin.ui.Button;
@@ -13,7 +13,7 @@ import com.vaadin.v7.data.fieldgroup.FieldGroup.CommitException;
 import com.vaadin.v7.ui.DateField;
 import com.vaadin.v7.ui.TextField;
 
-import my.vaadin.Undined.backend.Restaurant;
+import cs.dal.csci3130.Undined.backend.Restaurant;
 
 public class RequestForm extends FormLayout{
