+++ src/main/java/cs/dal/csci3130/Undined/backend/UserService.java
@@ -1,4 +1,4 @@
-package my.vaadin.Undined.backend;
+package cs.dal.csci3130.Undined.backend;
 
 import org.apache.commons.beanutils.BeanUtils;
 
@@ -8,9 +8,9 @@ import java.util.logging.Logger;
 
 // Backend service class. This is just a typical Java backend implementation
 // class and nothing Vaadin specific.
-public class MenuService {
+public class UserService {
 
-    private static MenuService instance;
+    private static UserService instance;
 
     //Code that originally filled the database with fake contacts
     /*
@@ -39,38 +39,38 @@ public class MenuService {
     }
 	*/
 
-    private HashMap<Long, Menu> menus = new HashMap<>();
+    private HashMap<Long, User> users = new HashMap<>();
     private long nextId = 0;
 
     //I think this is the search box.
-    public synchronized List<Menu> findAll(String stringFilter) {
-        ArrayList<Menu> arrayList = new ArrayList<Menu>();
+    public synchronized List<User> findAll(String stringFilter) {
+        ArrayList<User> arrayList = new ArrayList<User>();
         
-        //For each menu
-        for (Menu menu : menus.values()) {
+        //For each user
+        for (User user : users.values()) {
             try {
             	//If there is nothing in the search box
             	//or
-            	//The Menu.toString() contains the text in the search box (not case sensitive)
+            	//The User.toString() contains the text in the search box (not case sensitive)
                 boolean passesFilter = (stringFilter == null || stringFilter.isEmpty())
-                        || menu.toString().toLowerCase()
+                        || user.toString().toLowerCase()
                                 .contains(stringFilter.toLowerCase());
                 //Add it to the list that will be returned
                 if (passesFilter) {
-                    arrayList.add(menu.clone());
+                    arrayList.add(user.clone());
                 }
             
             //If error, log it
             } catch (CloneNotSupportedException ex) {
-                Logger.getLogger(MenuService.class.getName()).log(
+                Logger.getLogger(UserService.class.getName()).log(
                         Level.SEVERE, null, ex);
             }
         }
         
         //Sort the created list
-        Collections.sort(arrayList, new Comparator<Menu>() {
+        Collections.sort(arrayList, new Comparator<User>() {
             @Override
-            public int compare(Menu o1, Menu o2) {
+            public int compare(User o1, User o2) {
                 return (int) (o2.getId() - o1.getId());
             }
         });
@@ -80,23 +80,23 @@ public class MenuService {
     }
 
     public synchronized long count() {
-        return menus.size();
+        return users.size();
     }
 
-    public synchronized void delete(Menu value) {
-        menus.remove(value.getId());
+    public synchronized void delete(User value) {
+        users.remove(value.getId());
     }
 
-    public synchronized void save(Menu entry) {
+    public synchronized void save(User entry) {
         if (entry.getId() == null) {
             entry.setId(nextId++);
         }
         try {
-            entry = (Menu) BeanUtils.cloneBean(entry);
+            entry = (User) BeanUtils.cloneBean(entry);
         } catch (Exception ex) {
             throw new RuntimeException(ex);
         }
-        menus.put(entry.getId(), entry);
+        users.put(entry.getId(), entry);
     }
 
 }
